+++ react-components/src/actions.js
@@ -5,6 +5,7 @@ export const SET_INITIAL_STATE = 'SET_INITIAL_STATE'
 export const SET_PRODUCTS_EXPERTISE = 'SET_PRODUCTS_EXPERTISE'
 export const SET_COUNTRIES_EXPERTISE = 'SET_COUNTRIES_EXPERTISE'
 export const SET_PERFORM_FEATURE_SKIP_COOKIE_CHECK = 'SET_PERFORM_FEATURE_SKIP_COOKIE_CHECK'
+export const SET_NEXT_URL = 'SET_NEXT_URL'
 
 // action creators
 
@@ -47,10 +48,18 @@ const skipFeatureCookieCheck = function() {
   }
 }
 
+const setNextUrl = function(nextUrl) {
+  return {
+    type: SET_NEXT_URL,
+    payload: nextUrl,
+  }
+}
+
 export default {
   setInitialState,
   toggleModalIsOpen,
   setProductsExpertise,
   setCountriesExpertise,
   skipFeatureCookieCheck,
+  setNextUrl,
 }